Add string deserialization

+++ src/constants.rs
@@ -1,5 +1,5 @@
-#[cfg(feature = "no_std")]
+#[cfg(all(feature = "no_std", not(test)))]
 use core::u8::{MAX as MAX_U8_AS_U8};
-#[cfg(not(feature = "no_std"))]
+#[cfg(any(not(feature = "no_std"), test))]
 use std::u8::{MAX as MAX_U8_AS_U8};
-pub const MAX_U8: u64 = MAX_U8_AS_U8 as u64;
+pub const MAX_U8: u32 = MAX_U8_AS_U8 as u32;
